renamed linting result object filter related files to linting message filter
